+++ index.html
@@ -445,7 +445,6 @@
     </template>
 
     <!-- Подключаем сценарий script.js, используя относительный путь -->
-    <script src="./js/functions.js"></script>
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type="module"></script>
   </body>
 </html>
